Enhance README and implement transaction querying functionality. Added 'query_transaction' tool to retrieve transaction status and results by hash, including confirmation status and detailed results. Updated Cyberlink services to support real-time transaction monitoring and automatic status polling. Bump version to 0.2.1.
