+++ src/services/cyberlink/tx.service.ts
@@ -3,7 +3,7 @@ import { DirectSecp256k1HdWallet } from '@cosmjs/proto-signing';
 import { GasPrice, IndexedTx } from '@cosmjs/stargate';
 import { ErrorCode, McpError } from '@modelcontextprotocol/sdk/types.js';
 import { CyberlinkBaseService } from './base.service';
-import { TxCyberlinkResponseResult, TxResponse, TxStatusResponse } from './types';
+import { TxResponse } from './types';
 import { sanitizeQueryResult } from './utils';
 
 // Fields to extract from cyberlink event
@@ -52,72 +52,14 @@ export class CyberlinkTxService extends CyberlinkBaseService {
     }
   }
 
-  private ensureInitialized(): void {
+  protected ensureInitialized(): void {
     if (!this.signingClient || !this.wallet) {
       throw new McpError(ErrorCode.InternalError, 'CyberlinkTxService not initialized');
     }
   }
 
-  private async getTxStatus(transactionHash: string): Promise<TxStatusResponse> {
-    try {
-      this.ensureInitialized();
-      const tx = (await this.signingClient!.getTx(transactionHash)) as IndexedTx;
-
-      if (!tx) {
-        return { status: 'pending' };
-      }
-
-      const wasmEvent = tx.events.find((e) => e.type === 'wasm')?.attributes || [];
-
-      const result: Record<string, string> = {};
-
-      FIELD_NAMES.forEach((key) => {
-        const value = wasmEvent.find((a) => a.key === key)?.value;
-        if (value) {
-          result[key] = value;
-        }
-      });
-
-      if (tx.code === 0) {
-        return {
-          status: 'confirmed',
-          result: { ...result, transaction_hash: transactionHash },
-        };
-      }
-
-      return {
-        status: 'failed',
-        error: String(tx.code) || 'Transaction failed',
-      };
-    } catch (error) {
-      throw new McpError(ErrorCode.InternalError, 'Failed to query transaction', { error });
-    }
-  }
-
-  private async waitForTransaction(
-    txHash: string,
-    timeoutMs: number = 30000,
-    pollIntervalMs: number = 1000
-  ): Promise<TxCyberlinkResponseResult> {
-    const startTime = Date.now();
-
-    while (Date.now() - startTime < timeoutMs) {
-      try {
-        const result = await this.getTxStatus(txHash);
-
-        if (result.status === 'confirmed') {
-          return result.result || { transaction_hash: txHash };
-        } else if (result.status === 'failed') {
-          throw new Error(result.error || 'Transaction failed');
-        }
-      } catch (error) {
-        console.error(`Error while polling transaction ${txHash}:`, error);
-        // Continue polling if transaction not found
-      }
-      await new Promise((resolve) => setTimeout(resolve, pollIntervalMs));
-    }
-
-    throw new Error(`Transaction confirmation timed out after ${timeoutMs}ms`);
+  protected async getTx(transactionHash: string): Promise<IndexedTx | null> {
+    return this.signingClient!.getTx(transactionHash);
   }
 
   async executeTx(msg: any): Promise<TxResponse> {
@@ -148,7 +90,7 @@ export class CyberlinkTxService extends CyberlinkBaseService {
       return {
         status: 'completed',
         info: sanitizeQueryResult(txInfo),
-        result: sanitizeQueryResult(txResult || {}),
+        result: txResult,
       };
     } catch (error) {
       return {
